refs #22915 restructure: remove (seemingly) unnecessary Widget class
