--- scripts/Muon/GUI/ElementalAnalysis/PeriodicTable/periodic_table_widget.py
@@ -1,61 +1,0 @@
-from __future__ import absolute_import
-
-from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_view import PeriodicTableView
-from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_presenter import PeriodicTablePresenter
-from Muon.GUI.ElementalAnalysis.PeriodicTable.periodic_table_model import PeriodicTableModel
-
-from PyQt4 import QtGui
-
-
-class PeriodicTable(QtGui.QWidget):
-    def __init__(self, parent=None):
-        super(PeriodicTable, self).__init__(parent)
-        view = PeriodicTableView(parent)
-        model = PeriodicTableModel()
-        self._presenter = PeriodicTablePresenter(view, model)
-
-    @property
-    def presenter(self):
-        return self._presenter
-
-    @property
-    def widget(self):
-        return self._presenter.widget
-
-    @property
-    def peak_data(self):
-        return self._presenter.model.peak_data
-
-    @property
-    def selection(self):
-        return self.widget.ptable.getSelection()
-
-    def element_data(self, element):
-        try:
-            return self.presenter.model.peak_data[element]
-        except KeyError:
-            return None
-
-    def select_element(self, element, deselect=False):
-        self.widget.ptable.setElementSelected(element, not deselect)
-
-    def add_elements(self, *elements):
-        self.widget.ptable.setSelection(elements)
-
-    def register_table_changed(self, slot):
-        self._presenter.widget.sig_table_changed.connect(slot)
-
-    def unregister_table_changed(self, slot):
-        self._presenter.widget.sig_table_changed.disconnect(slot)
-
-    def register_table_lclicked(self, slot):
-        self._presenter.widget.sig_table_lclicked.connect(slot)
-
-    def unregister_table_lclicked(self, slot):
-        self._presenter.widget.sig_table_lclicked.disconnect(slot)
-
-    def register_table_rclicked(self, slot):
-        self._presenter.widget.sig_table_rclicked.connect(slot)
-
-    def unregister_table_rclicked(self, slot):
-        self._presenter.widget.sig_table_rclicked.disconnect(slot)
